requirements.txt
django>=5.2
pyyaml>=5.1
uritemplate>=3.0.0
django-filter>1.0.1
django-guardian>=1.1.1
markdown>=3.3.0
pygments>=2.7.0
djangorestframework>=3.16.0
uuid
python-dotenv
bcrypt
djangorestframework-simplejwt
django-cors-headers
django-extensions